Adding task 6

+++ 0x00-hello_world/6-size.c
@@ -7,9 +7,9 @@
 
 int main(void)
 {
-	printf("Size of a char: %d byte\n", sizeof(char));
-	printf("size of an int: %d byte\n", sizeof(int));
-	printf("Size of a float: %d byte\n", sizeof(float));
-	printf("Size of a long int: %d byte\n", sizeof(long int));
-	printf("Size of a long long int:%d byte\n", sizeof(long long int));
+	printf("Size of a char: %d byte\n", (int) sizeof(char));
+	printf("size of an int: %d byte\n", (int) sizeof(int));
+	printf("Size of a float: %d byte\n", (int) sizeof(float));
+	printf("Size of a long int: %d byte\n", (int) sizeof(long int));
+	printf("Size of a long long int:%d byte\n", (int) sizeof(long long int));
 }
